+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Removed
 - Remove channels 15 & 16 from the Channel enum as those aren't even valid
 - Remove ConnectionParameters impl
+- Remove irq pin and wake pin from Atwinc1500 struct
 
 ## [0.1.0] - 2022-10-13
 ### Added